+++ dxm/dxm.py
@@ -55,6 +55,7 @@ from dxm.lib.DxJobs.jobs_worker import job_delete
 from dxm.lib.DxJobs.jobs_worker import job_copy
 from dxm.lib.DxJobs.jobs_worker import job_update
 from dxm.lib.DxJobs.jobs_worker import job_cancel
+from dxm.lib.DxJobs.jobs_worker import jobs_report
 from dxm.lib.DxColumn.column_worker import column_list
 from dxm.lib.DxColumn.column_worker import column_setmasking
 from dxm.lib.DxColumn.column_worker import column_unsetmasking
@@ -1076,6 +1077,26 @@ def list(dxm_state, jobname, envname):
     """
     exit(jobs_list(dxm_state.engine, jobname, envname, dxm_state.format))
 
+@job.command()
+@click.option('--jobname', help="Filter jobs using jobname")
+@click.option('--envname', help='Filter jobs belongs to one environment')
+@click.option('--last', help='Display only last execution', is_flag=True)
+@click.option('--details', help='Display execution details', is_flag=True, default=False)
+@click.option('--startdate', help='Display jobs started after startdate', type=click.DateTime())
+@click.option('--enddate', help='Display jobs finished before enddate', type=click.DateTime())
+@common_options
+@pass_state
+def report(dxm_state, jobname, envname, last, startdate, enddate, details):
+    """
+    Display report of jobs defined in Masking Engine
+
+    If no filter options are specified, all jobs will be displayed.
+    If --envname or --jobname is set, output list will
+    be limited by value of this option and return non-zero return code
+    if jobname is not found.
+    """
+    exit(jobs_report(dxm_state.engine, jobname, envname, dxm_state.format, 
+                     last, startdate, enddate, details))
 
 @job.command()
 @click.option(